+++ src/main/java/com/nubiform/payment/vo/SentData.java
@@ -1,4 +1,4 @@
-package com.nubiform.payment.api.vo;
+package com.nubiform.payment.vo;
 
 import lombok.Data;
 import org.apache.commons.lang3.StringUtils;
