+++ src/main/java/com/nubiform/payment/vo/SubmitRequest.java
@@ -1,4 +1,4 @@
-package com.nubiform.payment.api.vo;
+package com.nubiform.payment.vo;
 
 import lombok.Data;
 import org.hibernate.validator.constraints.Range;
