+++ src/test/java/com/nubiform/payment/vo/CardTest.java
@@ -1,4 +1,4 @@
-package com.nubiform.payment.api.vo;
+package com.nubiform.payment.vo;
 
 import org.junit.jupiter.api.Test;
